+++ src/features/ChatInfo/index.tsx
@@ -1,12 +1,12 @@
 'use client';
 
-import { DraggablePanel } from '@lobehub/ui';
+import { DraggablePanel, TabsNav } from '@lobehub/ui';
 import { createStyles } from 'antd-style';
 import React, { useState } from 'react';
+import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
-import { CHAT_INFO_MAX_WIDTH, CHAT_INFO_WIDTH } from '@/constants/token';
-import Header from '@/features/ChatInfo/Header';
+import { CHAT_HEADER_HEIGHT, CHAT_INFO_MAX_WIDTH, CHAT_INFO_WIDTH } from '@/constants/token';
 import ChatList from '@/features/ChatList';
 import MotionList from '@/features/MotionList';
 import PlayList from '@/features/PlayList';
@@ -22,7 +22,11 @@ const useStyles = createStyles(({ css, token }) => ({
     flex-direction: column;
   `,
   header: css`
+    height: ${CHAT_HEADER_HEIGHT}px;
     border-bottom: 1px solid ${token.colorBorder};
+  `,
+  player: css`
+    min-width: 480px;
   `,
 }));
 
@@ -32,7 +36,8 @@ export default () => {
     s.setChatSidebar,
   ]);
 
-  const [tab, setTab] = useState<Tab>(Tab.History);
+  const [tab, setTab] = useState<Tab>(Tab.ChatList);
+  const { t } = useTranslation('features');
 
   const { styles } = useStyles();
 
@@ -48,9 +53,34 @@ export default () => {
       expand={showChatSidebar}
       placement={'right'}
     >
-      <Header tab={tab} setTab={setTab} />
+      <Flexbox justify={'space-between'} horizontal align={'center'} className={styles.header}>
+        <TabsNav
+          activeKey={tab}
+          items={[
+            {
+              label: t('info.chat'),
+              key: Tab.ChatList,
+            },
+            {
+              label: t('info.dance'),
+              key: Tab.PlayList,
+            },
+            {
+              label: t('info.motions'),
+              key: Tab.Motions,
+            },
+            {
+              label: t('info.posture'),
+              key: Tab.Posture,
+            },
+          ]}
+          onChange={(key) => {
+            setTab(key as Tab);
+          }}
+        />
+      </Flexbox>{' '}
       <Flexbox height={'calc(100vh - 128px)'}>
-        {tab === Tab.History && <ChatList />}
+        {tab === Tab.ChatList && <ChatList />}
         {tab === Tab.PlayList && <PlayList />}
         {tab === Tab.Motions && <MotionList />}
         {tab === Tab.Posture && <PostureList />}
