+++ src/features/ChatInput/Header/ActionBar/index.tsx
@@ -1,6 +1,6 @@
-import History from '@/features/Actions/History';
 import Record from '@/features/Actions/Record';
 import Token from '@/features/Actions/Token';
+import History from '@/features/ChatList/Actions/History';
 
 const ActionBar = () => (
   <>
