+++ src/features/ChatList/style.ts
@@ -1,6 +1,9 @@
 import { createStyles } from 'antd-style';
 
-export const useStyles = createStyles(({ css }) => ({
+export const useStyles = createStyles(({ css, token }) => ({
+  header: css`
+    padding: 0 ${token.paddingSM}px;
+  `,
   list: css`
     position: relative;
     height: 100%;
